+++ R/breaks.r
@@ -20,6 +20,24 @@ pretty_breaks <- function(n = 5, ...) {
   }
 }
 
+#' Produce breaks spaced by a fixed distance
+#'
+#' @param width Distance between each break
+#' @param offset Use if you don't want breaks to start at zero
+#' @export
+#' @examples
+#' demo_continuous(c(0, 100))
+#' demo_continuous(c(0, 100), breaks = breaks_width(10))
+#' demo_continuous(c(0, 100), breaks = breaks_width(20, -4))
+#' demo_continuous(c(0, 100), breaks = breaks_width(20, 4))
+breaks_width <- function(width, offset = 0) {
+  force_all(width, offset)
+
+  function(x) {
+    fullseq(x, width) + offset
+  }
+}
+
 #' Extended breaks
 #'
 #' Uses Wilkinson's extended breaks algorithm as implemented in the